One source menu, everywhere it shows up
The sidebar rows, gallery cards, and reader overflow each built their own
copy of the source menu; the graph and chat citations had none. Now one
builder (SourceMenu.tsx) produces the list — ask, refresh, edit, choose
card image, open original / show in Finder, copy, tags, note, file under
a card, remove — and one hook owns every modal those verbs open, so the
gallery card gains the image picker and the sidebar gains nothing it
has to maintain. Graph nodes get the menu on right-click (RowMenu learns
a host-driven open for non-DOM rows), and citation cards get the ⋯.

Not covered: the Registry's attached-document rows, which hold
cross-notebook attachments rather than Source objects.
